pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-windows-companion"
version = "0.1.0"
description = "Persistent Windows AI companion with overlay UI, Telegram alerts, and multi-backend chat"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyQt6>=6.6.0",
  "qasync>=0.27.1",
  "python-telegram-bot>=20.7",
  "psutil>=5.9.0",
  "wmi>=1.5.1",
  "aiosqlite>=0.19.0",
  "apscheduler>=3.10.0",
  "pyyaml>=6.0",
  "google-generativeai>=0.5.0",
  "groq>=0.9.0",
  "ollama>=0.1.8",
]

[project.scripts]
ai-companion = "main:main"

[tool.setuptools]
py-modules = ["main", "run"]
packages = ["ai_companion", "ai", "comms", "core", "integrations", "memory", "ui"]
